Move Flux specific implementation out of Statsd registry
The current implementation of the Statsd registry is tightly coupled to
the internal flux reactor event loop. This coupling makes it difficult
to plug in alternatives. This change intends to abstract the event loop
from the actual implementation of the event loop by hiding the flux
reactor loop behind the `MeterProcessor` interface and extending the builder
to allow advanced users to provide their own implementation.
