+++ implementations/micrometer-registry-statsd/src/main/java/io/micrometer/statsd/internal/MeterProcessor.java
@@ -1,0 +1,27 @@
+/*
+ * Copyright 2022 Spaceteams GmbH
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ * https://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package io.micrometer.statsd.internal;
+
+public interface MeterProcessor {
+
+    public void start(Runnable poll);
+
+    public void stop();
+
+    public void next(String line);
+
+}
