+++ implementations/micrometer-registry-statsd/src/main/java/io/micrometer/statsd/internal/flux/LogbackMetricsSuppressingFluxSink.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package io.micrometer.statsd.internal;
+package io.micrometer.statsd.internal.flux;
 
 import io.micrometer.core.instrument.binder.logging.LogbackMetrics;
 import reactor.core.Disposable;
