+++ implementations/micrometer-registry-statsd/src/test/java/io/micrometer/statsd/internal/flux/BufferingFluxTest.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package io.micrometer.statsd.internal;
+package io.micrometer.statsd.internal.flux;
 
 import org.junit.jupiter.api.Test;
 import reactor.core.publisher.Flux;
